docs: DTO 필드 변환 및 test 문서구조 변경

+++ src/main/java/com/bankersCalculator/server/advice/loanAdvice/dto/LoanAdviceRequest.java
@@ -53,7 +53,7 @@ public class LoanAdviceRequest {
     // 주택정보
     private RentHousingType rentHousingType;  // 주택타입
 
-    private Long exclusiveArea; // 전용면적
+    private Double exclusiveArea; // 전용면적
 
     private String buildingName;   // 건물명
